% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R, R/grafico_antigo.R
\name{grafico}
\alias{grafico}
\title{Gráficos de verificação do modelo linear}
\usage{
grafico(modelo, tipo)

grafico(modelo, tipo)
}
\arguments{
\item{modelo}{Modelo de regressão linear simples ou múltipla. Deve ser, obrigatoriamente, da classe \code{"modelo_linear"}.
Deve conter os elementos \code{dados}, \code{residuos} (ou seja, os resíduos do modelo) e \code{ajustados} (ou seja, os valores ajustados (preditos) pelo modelo).
O modelo fornecido deve ter a classe "modelo_linear", caso contrário, a função gerará um erro.}

\item{tipo}{É uma string que especifica o tipo de gráfico desejado. Os valores aceitos são:
\describe{
\item{rvp}{Gera gráficos de Resíduos vs Preditoras.}
\item{rva}{Gera um gráfico de Resíduos vs Valores Ajustados.}
\item{normres}{Gera um Gráfico QQPlot para avaliar a normalidade dos resíduos.}
}}
}
\description{
A função \code{grafico} é utilizada para criação de gráficos para verificação do modelo linear resultante. Válida tanto para regressão linear simples quanto para regressão linear multivariada. Ela permite visualizar o ajuste do modelo e verificar os pressupostos dos resíduos.

\code{grafico} Função utilizada para criação de gráficos de resíduos em regressão linear. Válida também para regressão linear múltipla.
}
\examples{
# Carregando o conjunto de dados do pacote
data(rl_dataset)

# Ajustando o modelo de regressão linear com o rl_dataset
modelo_exemplo <- reg_linear(rl_dataset[, c("Preditora.1", "Preditora.2", "Preditora.3", "Preditora.4")],
                             rl_dataset[, c("Resposta.1", "Resposta.2")])

# Gerando gráficos de verificação do modelo ajustado
# Gráfico de Valores Preditos x Observados
grafico(modelo_exemplo, tipo = "pvo")

# Gráfico de Resíduos x Preditoras
grafico(modelo_exemplo, tipo = "rvp")

# QQ-plot dos Resíduos
grafico(modelo_exemplo, tipo = "qqplot")

grafico(modelo, "rvp")      # Gráfico de Resíduos vs Preditoras
grafico(modelo, "rva")      # Gráfico de Resíduos vs Valores Ajustados
grafico(modelo, "normres")  # Gráfico Q-Q de Normalidade dos Resíduos



}
