% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicao}
\alias{predicao}
\title{Função de Predição para Modelos Lineares}
\usage{
predicao(modelo, preditores)
}
\arguments{
\item{modelo}{Um modelo de regressão linear da classe \code{"modelo_linear"}. Caso contrário, a função irá gerar um erro.}

\item{preditores}{Uma matriz ou data frame contendo os valores preditores. O número de preditores deve ser compatível com o modelo ajustado.}
}
\value{
Retorna um objeto da classe \code{"predicao"}, contendo:
\item{preditores}{Os preditores utilizados na predição (incluindo a coluna de intercepto).}
\item{preditos}{Os valores preditos calculados pelo modelo.}
}
\description{
A função \code{predicao} realiza predições a partir de um modelo de regressão linear ajustado. Ela utiliza o modelo fornecido e aplica os preditores fornecidos para calcular os valores preditos.
A função pode ser aplicada tanto em modelos de regressão simples quanto multivariada.
}
\examples{
# Carregando o conjunto de dados do pacote
data(rl_dataset)

# Ajustando o modelo de regressão linear com o rl_dataset
modelo_exemplo <- reg_linear(rl_dataset[, c("Preditora.1", "Preditora.2", "Preditora.3", "Preditora.4")],
                             rl_dataset[, c("Resposta.1", "Resposta.2")])

# Fazendo predições com novos valores de preditores
novos_preditores <- matrix(c(3, 4, 5, 2, 1, 6, 7, 3), nrow = 2, byrow = TRUE)
resultado_predicao <- predicao(modelo_exemplo, novos_preditores)


resultado_predicao$preditos

}
