% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_linear.R
\name{reg_linear}
\alias{reg_linear}
\title{Ajuste de Modelos Lineares}
\usage{
reg_linear(x, y)
}
\arguments{
\item{x}{Variáveis preditores do modelo. Deve ser um objeto do tipo \code{numeric} que seja coercível a uma matrix pela função \code{as.matrix}.}

\item{y}{Variáveis respostas do modelo. Deve ser um objeto do tipo \code{numeric} que seja coercível a uma matrix pela função \code{as.matrix}.}
}
\value{
\code{reg_linear} retorna uma lista da classe "modelo_linear". Essa lista possui os seguintes dados:
\describe{
  \item{coeficientes}{os valores estimados dos coeficientes do modelo.}
  \item{dados}{os dados utilizados para se criar o modelo.}
  \item{residuos}{os resíduos; ou seja, o valor da resposta menos o valor ajustado}
  \item{ajustado}{os valores de ajuste do modelo; ou seja, os valores de \code{x} aplicados à função de regressão.}
}

FALTA EXEMPLO
}
\description{
\code{reg_linear} deve ser usada pra criar modelo de regressão lineares, incluindo multivariados.
}
