% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLinearME918.R
\docType{package}
\name{RegLinearME918-package}
\alias{RegLinearME918}
\alias{RegLinearME918-package}
\title{RegLinearME918: Um Pacote para Regressão Linear Simples}
\description{
O pacote RegLinearME918 oferece ferramentas para realizar regressão linear simples, fazer predições, e gerar gráficos com base nos modelos ajustados.
}
\section{Funções principais}{

\itemize{
\item \code{\link{reg_linear}}: Ajusta um modelo de regressão linear.
\item \code{\link{predicao}}: Faz predições com base no modelo ajustado.
\item \code{\link{grafico}}: Gera um gráfico comparando os valores observados e preditos.
}
}

\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Luiz Felipe de Oliveira Barbosa Nunes \email{l255403@dac.unicamp.br}
  \item Enzo Putton Tortelli \email{e204256@dac.unicamp.br}
  \item Lara Maria Herrera Drugowick \email{l236181@dac.unicamp.br}
  \item Eric Pavarim Lima \email{e247471@dac.unicamp.br}
}

}
\keyword{internal}
