% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLinearME918.R
\docType{package}
\name{RegLinearME918}
\alias{RegLinearME918-package}
\alias{RegLinearME918}
\title{RegLinearME918: Um Pacote para Regressão Linear Simples e Múltipla}
\description{
O pacote RegLinearME918 oferece ferramentas para ajustar modelos de regressão linear simples e múltipla, fazer predições com base nos modelos ajustados e gerar gráficos que ajudam a visualizar e analisar os resultados.
}
\section{Funções principais}{

\itemize{
\item \code{\link{reg_linear}}: Ajusta um modelo de regressão linear com base em variáveis preditoras e respostas.
\item \code{\link{predicao}}: Realiza predições com base no modelo de regressão ajustado e novos dados preditores.
\item \code{\link{grafico}}: Gera gráficos para visualizar comparações entre valores observados e preditos, resíduos e QQ-plots dos resíduos.
}
}

\section{Conjunto de dados}{

O pacote inclui um conjunto de dados simulado, \code{\link{rl_dataset}}, que pode ser utilizado para testar as funções do pacote.
}

\seealso{
\code{\link{reg_linear}}, \code{\link{predicao}}, \code{\link{grafico}}, \code{\link{rl_dataset}}
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Luiz Felipe de Oliveira Barbosa Nunes \email{l255403@dac.unicamp.br}
  \item Enzo Putton Tortelli \email{e204256@dac.unicamp.br}
  \item Lara Maria Herrera Drugowick \email{l236181@dac.unicamp.br}
  \item Eric Pavarim Lima \email{e247471@dac.unicamp.br}
}

}
\keyword{dados}
\keyword{estatísticos,}
\keyword{gráficos,}
\keyword{linear,}
\keyword{modelos}
\keyword{predição,}
\keyword{regressão}
\keyword{simulados}
