% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{grafico}
\alias{grafico}
\title{Gráficos de verificação do modelo linear}
\usage{
grafico(modelo, tipo = c("pvo", "rvp", "qqplot"))
}
\arguments{
\item{modelo}{Modelo de regressão linear. Deve ser, obrigatoriamente, da classe \code{"modelo_linear"}. Caso contrário, a função gerará um erro.}

\item{tipo}{Vetor do tipo \code{char} que especifica os tipos de gráficos desejados. Os valores aceitos são:
\describe{
\item{pvo}{Gráfico de valores preditos x valores observados.}
\item{rvp}{Gráfico de resíduos x preditora.}
\item{qqplot}{QQ-plot dos resíduos.}
}}
}
\description{
A função \code{grafico} é utilizada para criação de gráficos para verificação do modelo linear resultante. Válida tanto para regressão linear simples quanto para regressão linear multivariada. Ela permite visualizar o ajuste do modelo e verificar os pressupostos dos resíduos.
}
\examples{
# Carregando o conjunto de dados do pacote
data(rl_dataset)

# Ajustando o modelo de regressão linear com o rl_dataset
modelo_exemplo <- reg_linear(rl_dataset[, 3:5], rl_dataset[, 1:2])

# Gerando gráficos de verificação do modelo ajustado
grafico(modelo_exemplo)

}
