% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rl_dataset}
\alias{rl_dataset}
\title{Desempenho de computadores}
\format{
Objeto \code{data.frame} com cinco colunas:
\describe{
\item{price}{Preço do computador (em dólares americanos);}
\item{boot_speed}{Tempo para que o computador inicialize o sistema operacional (em segundos);}
\item{gpu_benchmark}{Pontução do computador no benchmark PavarimMark3D™;}
\item{cpu_benchmark}{Pontuação do computador no benchmark PavarimMark™ }
\item{ram_qty}{Quantidade memória RAM no computador (em GB).}
}
}
\source{
Gerado artificialmente para fins de exemplo no pacote.
}
\usage{
rl_dataset
}
\description{
Este é um conjunto de dados de exemplo incluído no pacote, gerado artificialmente para fins ilustrativos.
}
\keyword{datasets}
