% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meu_dataset.R
\docType{data}
\name{rl_dataset}
\alias{rl_dataset}
\title{Conjunto de Dados Simulados}
\format{
Objeto \code{data.frame} com 1000 observações e 4 variáveis:

\describe{
  \item{Preditora 1}{Variável explicativa gerada a partir de uma distribuição normal com média 3 e desvio-padrão 1.}
  \item{Preditora 2}{Variável explicativa gerada a partir de uma distribuição Poisson com parâmetro lambda = 5.}
  \item{Preditora 3}{Variável explicativa gerada a partir de uma distribuição normal com média 4 e desvio-padrão 1.}
  \item{Resposta}{Variável resposta gerada pela fórmula: \code{Resposta = 0.8 * Preditora 1 + 3 * Preditora 2 + 1.5 * Preditora 3 + 3.6 + rnorm(1000, 0, 1)}.}
}
}
\source{
Dados gerados através de simulações.
}
\usage{
meu_dataset
}
\description{
Conjunto de dados criado para ser utilizado em funções de regressão linear múltipla desenvolvidas para o pacote.
}
\examples{
data(rl_dataset)
head(rl_dataset)
}
\keyword{datasets}
